fix(appearance): Density radio group was unresponsive
Changed the way the update source the density value to retrieve it
directly from the underlying config struct. CosmicTK currently has an
issue where changes aren't propagated back to the wrapper which cause
the UI to never be able to change the selected radio button on the radio
group for Density.

This is most likely a temporary fix until a fix upstream in libcosmic is
made.

The Config struct becomes stale when the underlying store has values
changed. Getter method like the one for the interface density only ever returns
the value it was initiated with:

https://github.com/pop-os/libcosmic/blob/944c6761f73f097552042912780104783f9ca62e/src/config/mod.rs#L70